mitosheet: make multi column resize work
